server/pyproject.toml
[tool.poetry]
name = "pyquizserver"
version = "0.1.0"
description = ""
authors = ["Igor Tolmachov <me@igorek.dev>"]
license = "GPL-3.0-only"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.10.2"
fastapi = "^0.85.1"
uvicorn = "^0.18.3"
SQLAlchemy = "^2b2"
psycopg2 = "^2.9.4"
PyJWT = "^2.6.0"

[tool.poetry.group.dev.dependencies]
mypy = "^0.982"
black = "^22.10.0"
isort = "^5.10.1"

[tool.black]
line-length = 90

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_untyped_calls = true

[tool.isort]
profile = "black"
filter_files = true
extend_skip = ["__init__.py"]